refactor(chat): separate logic into hooks
