+++ frontend/app/chat/helpers/isSpeechRecognitionSupported.ts
@@ -1,6 +1,6 @@
 export function isSpeechRecognitionSupported() {
   if (
-    typeof window !== undefined &&
+    typeof window !== "undefined" &&
     ("SpeechRecognition" in window || "webkitSpeechRecognition" in window)
   ) {
     return true;
